+++ Models/Product.cs
@@ -8,25 +8,25 @@ namespace WebsiteBanCaPhe.Models
         [Key]
         public int ProductId { get; set; }
 
-        [Required] 
+        [Required]
         public string? ProductName { get; set; }
 
-        [Required] 
+        [Required]
         public string? ProductDescription { get; set; }
 
-        [Required] 
+        [Required]
         public string? PhotoURL { get; set; }
 
-        [Required] 
+        [Required]
         public string? Origin { get; set; }
 
-        [Required] 
+        [Required]
         public string? Branch { get; set; }
 
-        [Required] 
-        public decimal Price { get; set; }
+        [Required]
+        public long Price { get; set; }
 
-        [Required] 
+        [Required]
         public int Quantity { get; set; }
 
         public int QuantitySold { get; set; } = 0;
